+++ src/main/kotlin/models/profile/AchievementDescription.kt
@@ -1,7 +1,8 @@
 package models.profile
 
 data class AchievementDescription(
-    val place: String,
+    val type: String,
+    val title: String,
     val description: String,
     val date: String
 )
